requirements.txt
streamlit==1.25.0
openai==0.27.8
pytesseract==0.3.10
pdf2image==1.16.3
PyPDF2==3.0.1
python-dotenv==1.0.0
pyzerox==0.1.2
